chore: update deps, rustc, tests
Update Rust version to 1.85.
Update all deps to latest.
Fix failing test on main.

The failing test is due to the relaxed atomic ordering. I don't
want to use a stricter ordering and some out-of-ordering is okay.
The test was fixed by fixing the number of threads to 2 and allow-
ing messages to be out of order by +- 10.